backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djoser==2.2.2
django-filter==23.4
Pillow==10.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
drf-extra-fields==3.7.0
reportlab==4.0.8
django-cors-headers==4.3.1 